refactor: Split storage route into separate files
- Create helpers.ts for R2 and Google Drive functions
- Create post.ts for file upload handler
- Create delete.ts for file deletion handler
- Refactor route.ts to import and integrate methods
- Improve code organization and maintainability
